added loading indicator
