+++ src/components/RestaurantList.js
@@ -1,21 +1,24 @@
-import { List, ListItem, ListItemText } from '@mui/material';
+import { CircularProgress, List, ListItem, ListItemText } from '@mui/material';
 import { useEffect } from 'react';
 import { connect } from 'react-redux';
 import { loadRestaurants } from '../store/restaurants/actions';
 
-export function RestaurantList({ loadRestaurants, restaurants }) {
+export function RestaurantList({ loadRestaurants, restaurants, loading }) {
   useEffect(() => {
     loadRestaurants();
   }, [loadRestaurants]);
 
   return (
-    <List>
-      {restaurants.map((restaurant) => (
-        <ListItem key={restaurant.id}>
-          <ListItemText>{restaurant.name}</ListItemText>
-        </ListItem>
-      ))}
-    </List>
+    <>
+      {loading && <CircularProgress />}
+      <List>
+        {restaurants.map((restaurant) => (
+          <ListItem key={restaurant.id}>
+            <ListItemText>{restaurant.name}</ListItemText>
+          </ListItem>
+        ))}
+      </List>
+    </>
   );
 }
